Notes about the Shapefile size warning
Blake, I was originally confused about linking to the Shapefile Warn section from the store. On reflection I think the point I added in the list here + a suggestion to use GeoPKG are what I was looking for as someone coming from the store. Please edit at will of course, just suggestions.